Update patch set 4
Patch Set 4:

Yep, you can do that, but you'll need the whole sources to test anyway.

For any repository, clone it, then grab the appropriate gerrit project URL for it (ssh://<username>@r.cyanogenmod.com:29418/<project>) for CM, add it as a remote, and then push to refs/for/<branch> to that remote.

Patch-set: 4